Rename pkg to finder
